making a binary geo transform base that optimizes out the literal transforms
