+++ pinot-core/src/main/java/org/apache/pinot/core/geospatial/transform/function/StWithinFunction.java
@@ -18,18 +18,8 @@
  */
 package org.apache.pinot.core.geospatial.transform.function;
 
-import com.google.common.base.Preconditions;
-import java.util.List;
-import java.util.Map;
-import org.apache.pinot.core.operator.blocks.ProjectionBlock;
 import org.apache.pinot.core.operator.transform.TransformResultMetadata;
-import org.apache.pinot.core.operator.transform.function.BaseTransformFunction;
-import org.apache.pinot.core.operator.transform.function.LiteralTransformFunction;
-import org.apache.pinot.core.operator.transform.function.TransformFunction;
-import org.apache.pinot.segment.local.utils.GeometrySerializer;
 import org.apache.pinot.segment.local.utils.GeometryUtils;
-import org.apache.pinot.segment.spi.datasource.DataSource;
-import org.apache.pinot.spi.data.FieldSpec;
 import org.locationtech.jts.geom.Geometry;
 
 
@@ -37,37 +27,12 @@ import org.locationtech.jts.geom.Geometry;
  * Function that checks the containment of the two geo-spatial objects. It returns true if and only if
  * first geometry is completely inside second geometry.
  */
-public class StWithinFunction extends BaseTransformFunction {
+public class StWithinFunction extends BaseBinaryGeoTransformFunction {
   public static final String FUNCTION_NAME = "ST_Within";
-  private TransformFunction _firstArgument;
-  private TransformFunction _secondArgument;
-  private int[] _results;
 
   @Override
   public String getName() {
     return FUNCTION_NAME;
-  }
-
-  @Override
-  public void init(List<TransformFunction> arguments, Map<String, DataSource> dataSourceMap) {
-    Preconditions
-        .checkArgument(arguments.size() == 2, "2 arguments are required for transform function: %s", getName());
-    TransformFunction transformFunction = arguments.get(0);
-    Preconditions.checkArgument(transformFunction.getResultMetadata().isSingleValue(),
-        "First argument must be single-valued for transform function: %s", getName());
-    Preconditions.checkArgument(transformFunction.getResultMetadata().getDataType() == FieldSpec.DataType.BYTES
-        || transformFunction instanceof LiteralTransformFunction,
-        "The first argument must be of type BYTES, but was %s",
-        transformFunction.getResultMetadata().getDataType());
-    _firstArgument = transformFunction;
-    transformFunction = arguments.get(1);
-    Preconditions.checkArgument(transformFunction.getResultMetadata().isSingleValue(),
-        "Second argument must be single-valued for transform function: %s", getName());
-    Preconditions.checkArgument(transformFunction.getResultMetadata().getDataType() == FieldSpec.DataType.BYTES
-        || transformFunction instanceof LiteralTransformFunction,
-        "The second argument must be of types BYTES, but was %s",
-        transformFunction.getResultMetadata().getDataType());
-    _secondArgument = transformFunction;
   }
 
   @Override
@@ -76,20 +41,10 @@ public class StWithinFunction extends BaseTransformFunction {
   }
 
   @Override
-  public int[] transformToIntValuesSV(ProjectionBlock projectionBlock) {
-    if (_results == null) {
-      _results = new int[projectionBlock.getNumDocs()];
+  public int transformGeometryToInt(Geometry firstGeometry, Geometry secondGeometry) {
+    if (GeometryUtils.isGeography(firstGeometry) || GeometryUtils.isGeography(secondGeometry)) {
+      throw new RuntimeException(String.format("%s is available for Geometry objects only", FUNCTION_NAME));
     }
-    byte[][] firstValues = _firstArgument.transformToBytesValuesSV(projectionBlock);
-    byte[][] secondValues = _secondArgument.transformToBytesValuesSV(projectionBlock);
-    for (int i = 0; i < projectionBlock.getNumDocs(); i++) {
-      Geometry firstGeometry = GeometrySerializer.deserialize(firstValues[i]);
-      Geometry secondGeometry = GeometrySerializer.deserialize(secondValues[i]);
-      if (GeometryUtils.isGeography(firstGeometry) || GeometryUtils.isGeography(secondGeometry)) {
-        throw new RuntimeException(String.format("%s is available for Geometry objects only", FUNCTION_NAME));
-      }
-      _results[i] = firstGeometry.within(secondGeometry) ? 1 : 0;
-    }
-    return _results;
+    return firstGeometry.within(secondGeometry) ? 1 : 0;
   }
 }
